refactor: group messaging APIs
closes: #6854
