+++ messaging/api_sender.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"crypto/ecdsa"
 
+	"github.com/status-im/status-go/messaging/layers/transport"
 	"github.com/status-im/status-go/messaging/types"
 )
 
@@ -17,4 +18,12 @@ func (a *API) SendPrivate(ctx context.Context, params types.SendPrivateParams) e
 
 func (a *API) SendPrivateHashRatchetKeys(ctx context.Context, recipients []*ecdsa.PublicKey, groupID []byte) error {
 	return a.core.controller.Sender().SendPrivateHashRatchetKeys(ctx, recipients, groupID)
+}
+
+func (a *API) GetEphemeralKey() (*ecdsa.PrivateKey, error) {
+	return a.core.controller.Processor().GetEphemeralKey()
+}
+
+func ContactCodeTopic(publicKey *ecdsa.PublicKey) string {
+	return transport.ContactCodeTopic(publicKey)
 }
